requirements.txt
psutil>=5.9.8
pycaw>=20240210
python-i18n>=0.3.9
pywin32>=306
requests>=2.31.0